Fix bug in R/pipefittr.R
Fixes a bug that caused something like

```put_a_ring_on_it(filter(load('single_ladies.RData'), like_it))```

to fail where something like

```join(select(box, left), everything_you_own)```

would work.

This happened  because we weren't checking the length of the arguments
list before 'popping' out the first item. When the length was 1, as in
the case of `load('single_ladies.RData')` in the example above, then
the result was an empty vector of length 0.

Now, if the length of the arguments list is 1, we return a vector of
length 1, with an empty string in it.